// src/api/models/OverviewSchema.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Jingle Jam Community Page Public API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { OverviewSchemaCollections } from './OverviewSchemaCollections';
import {
    OverviewSchemaCollectionsFromJSON,
    OverviewSchemaCollectionsFromJSONTyped,
    OverviewSchemaCollectionsToJSON,
} from './OverviewSchemaCollections';
import type { JJRaised } from './JJRaised';
import {
    JJRaisedFromJSON,
    JJRaisedFromJSONTyped,
    JJRaisedToJSON,
} from './JJRaised';

/**
 * 
 * @export
 * @interface OverviewSchema
 */
export interface OverviewSchema {
    /**
     * 
     * @type {JJRaised}
     * @memberof OverviewSchema
     */
    raised: JJRaised;
    /**
     * 
     * @type {OverviewSchemaCollections}
     * @memberof OverviewSchema
     */
    collections: OverviewSchemaCollections;
    /**
     * 
     * @type {number}
     * @memberof OverviewSchema
     */
    donations: number;
    /**
     * 
     * @type {Date}
     * @memberof OverviewSchema
     */
    date: Date;
}

/**
 * Check if a given object implements the OverviewSchema interface.
 */
export function instanceOfOverviewSchema(value: object): value is OverviewSchema {
    if (!('raised' in value) || value['raised'] === undefined) return false;
    if (!('collections' in value) || value['collections'] === undefined) return false;
    if (!('donations' in value) || value['donations'] === undefined) return false;
    if (!('date' in value) || value['date'] === undefined) return false;
    return true;
}

export function OverviewSchemaFromJSON(json: any): OverviewSchema {
    return OverviewSchemaFromJSONTyped(json, false);
}

export function OverviewSchemaFromJSONTyped(json: any, ignoreDiscriminator: boolean): OverviewSchema {
    if (json == null) {
        return json;
    }
    return {
        
        'raised': JJRaisedFromJSON(json['raised']),
        'collections': OverviewSchemaCollectionsFromJSON(json['collections']),
        'donations': json['donations'],
        'date': (new Date(json['date'])),
    };
}

export function OverviewSchemaToJSON(value?: OverviewSchema | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'raised': JJRaisedToJSON(value['raised']),
        'collections': OverviewSchemaCollectionsToJSON(value['collections']),
        'donations': value['donations'],
        'date': ((value['date']).toISOString()),
    };
}